Phase 5 Batch 1b: drop the I prefix from static classes
A leading I on a static class reads as an interface. All remaining
I-prefixed extension classes are renamed to the same
Asm<Package><Target>Extensions convention used in Batch 1a — uniform
and collision-proof (a naive drop-the-I would have re-created CS0433 for
EndpointRouteBuilderExtensions and QueryableExtensions, each defined in
multiple assemblies under a shared namespace). Namespaces are unchanged,
so only by-name references need updating; extension-method call sites do
not. IPAddressExtensions keeps its IP (part of the type name) and
becomes AsmNetIPAddressExtensions.

IIdentifiableEqualityComparer is a real instantiable class, so it is
renamed to IdentifiableEqualityComparer and ships with an [Obsolete]
forwarding alias for one major cycle.
