+++ src/Asm.AspNetCore/Reporting/AsmAspNetCoreEndpointRouteBuilderExtensions.cs
@@ -10,7 +10,7 @@ namespace Microsoft.AspNetCore.Routing;
 /// <summary>
 /// Endpoint-routing extensions for mapping the security-reporting endpoints.
 /// </summary>
-public static class IEndpointRouteBuilderExtensions
+public static class AsmAspNetCoreEndpointRouteBuilderExtensions
 {
     /// <summary>Logger category used for integrity-report messages.</summary>
     public const string IntegrityLoggerCategory = "Asm.AspNetCore.Reporting.Integrity";
